% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heap_sort.R
\name{build_heap_recursive}
\alias{build_heap_recursive}
\title{Build Heap Recursively for Heap Sort}
\usage{
build_heap_recursive(heap, i, n)
}
\arguments{
\item{heap}{The heap to be built.}

\item{i}{The current index to start the heapification.}

\item{n}{The size of the heap.}
}
\value{
The heap after heapification.
}
\description{
This function builds a max heap recursively.
}
