% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radix_sort.R
\name{counting_sort}
\alias{counting_sort}
\title{Counting Sort as a Subroutine for Radix Sort}
\usage{
counting_sort(vec, exp)
}
\arguments{
\item{vec}{A numeric vector to be sorted.}

\item{exp}{The current digit place (1s, 10s, 100s, etc.).}
}
\value{
A vector sorted by the current digit.
}
\description{
This function performs counting sort on a specific digit (specified by 'exp').
}
