% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_Algorithms.R
\name{heap_sort}
\alias{heap_sort}
\title{Heap Sort Algorithm}
\usage{
heap_sort(v, type = "notRecursive")
}
\arguments{
\item{v}{A numeric vector to be sorted.}

\item{type}{Type of heap sort ('notRecursive' or 'recursive').}
}
\value{
A sorted numeric vector.
}
\description{
This function sorts a numeric vector using the Heap Sort algorithm.
It can either use an iterative or recursive approach.
}
