% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_Algorithms.R
\name{merge}
\alias{merge}
\title{Merge Two Sorted Vectors}
\usage{
merge(left, right)
}
\arguments{
\item{left}{A sorted numeric vector.}

\item{right}{A sorted numeric vector.}
}
\value{
A merged and sorted numeric vector.
}
\description{
This function merges two sorted vectors into a single sorted vector.
}
