% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sort.R
\name{merge_sort}
\alias{merge_sort}
\title{Merge Sort Algorithm}
\usage{
merge_sort(vec)
}
\arguments{
\item{vec}{A numeric vector to be sorted.}
}
\value{
A sorted numeric vector.
}
\description{
This function sorts a numeric vector using the Merge Sort algorithm.
It recursively divides the vector into sub-vectors and merges them back together.
}
