% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_sort.R
\name{quick_sort}
\alias{quick_sort}
\title{Quick Sort Algorithm}
\usage{
quick_sort(v)
}
\arguments{
\item{v}{A numeric vector to be sorted.}
}
\value{
A sorted numeric vector.
}
\description{
This function sorts a numeric vector using the Quick Sort algorithm.
It recursively divides the vector into sub-vectors and sorts each.
}
