% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_Algorithms.R
\name{radix_sort}
\alias{radix_sort}
\title{Radix Sort Algorithm}
\usage{
radix_sort(vec)
}
\arguments{
\item{vec}{A numeric vector to be sorted.}
}
\value{
A sorted numeric vector.
}
\description{
This function sorts a numeric vector using the Radix Sort algorithm.
The function iterates over each digit of the numbers and uses counting sort as a subroutine.
}
