% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_algorithms_Rcpp.R
\name{radix_sort_Rcpp}
\alias{radix_sort_Rcpp}
\title{Tri par tas en C++}
\arguments{
\item{x}{Un vecteur numérique à trier.}
}
\value{
Un vecteur trié.
}
\description{
Trie un vecteur numérique en utilisant l'algorithme de tri par tas (heap sort).
}
